+++ package.json
@@ -103,7 +103,7 @@
     "ipfs-unixfs-engine": "^0.14.2",
     "ipld-resolver": "^0.4.1",
     "isstream": "^0.1.2",
-    "libp2p-floodsub": "0.6.0",
+    "libp2p-floodsub": "0.7.0",
     "joi": "^10.0.6",
     "libp2p-ipfs-nodejs": "^0.17.1",
     "libp2p-ipfs-browser": "^0.17.3",
