+++ src/core/components/go-online.js
@@ -5,13 +5,13 @@ const Bitswap = require('ipfs-bitswap')
 const FloodSub = require('libp2p-floodsub')
 
 module.exports = (self) => {
-  return (cb) => {
+  return (callback) => {
     series([
       self.load,
       self.libp2p.start
     ], (err) => {
       if (err) {
-        return cb(err)
+        return callback(err)
       }
 
       self._bitswap = new Bitswap(
@@ -20,10 +20,11 @@ module.exports = (self) => {
         self._libp2pNode.peerBook
       )
       self._bitswap.start()
+
       self._blockService.goOnline(self._bitswap)
+
       self._pubsub = new FloodSub(self._libp2pNode)
-      // TODO self.pubsub.start()
-      cb()
+      self._pubsub.start(callback)
     })
   }
 }
