--- test/core/node-only/test-swarm-2.js
@@ -1,20 +1,0 @@
-/* eslint-env mocha */
-
-'use strict'
-
-const test = require('interface-ipfs-core')
-const IPFSFactory = require('../../utils/factory-core')
-
-let factory
-
-const common = {
-  setup: function (cb) {
-    factory = new IPFSFactory()
-    cb(null, factory)
-  },
-  teardown: function (cb) {
-    factory.dismantle(cb)
-  }
-}
-
-test.swarm(common)
